+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# v0.0.2 - 2018-10-16
+- Allow status bar button to be hidden
+- Add tooltip to status bar button
+
 # v0.0.1 - 2018-10-16
 - Initial release